Move tests into packages referencing shared validate
