+++ internal/storage/inmemory/user_session_repository_test.go
@@ -1,0 +1,11 @@
+package inmemory
+
+import (
+	"testing"
+
+	"github.com/willbicks/epigram/internal/storage/validate"
+)
+
+func TestUserSessionRepository(t *testing.T) {
+	validate.UserSessionRepository(t, NewUserSessionRepository())
+}
